+++ evealert/__init__.py
@@ -1,4 +1,4 @@
 """Initialize the app"""
 
-__version__ = "6.3.32"
+__version__ = "6.3.33"
 __title__ = "EVE Alert"